requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
google-auth==2.30.0
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
